requirements.txt
# Core packages for SAS protocol and serial communication
pyserial==3.5
crccheck==1.3.0
netifaces==0.11.0

# System utilities
psutil>=5.8
distro>=1.6  # Linux only



# Optional: Development and debugging tools
# requests>=2.28  # For HTTP requests if needed
# python-dateutil>=2.8  # For date handling 

# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
websockets==12.0

# Async and threading
# (No additional async dependencies needed beyond FastAPI's built-in support)

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Configuration and environment
python-dotenv==1.0.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Logging and monitoring
structlog==23.2.0 